replace v8-profiler with node.js v7.x compatible fork
v8-profiler had to be removed because it is not compatible with node.js
v7.x: `npm install` fails with compilation errors of C source code.
[0][1]

risingstack fork includes a fix that has not been merged into
node-inspector/v8-profiler yet. This is still compatible with Node.js
v6.x.

[0] node-inspector/v8-profiler#98
[1] node-inspector/v8-profiler#99